Add routed navigation and shared Zustand state
