+++ apps/desktop/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@base-ui/react": "^1.7.0",
     "@fontsource-variable/geist": "^5.3.0",
+    "@tanstack/react-router": "^1.170.31",
     "@tauri-apps/api": "^2.8.0",
     "@tauri-apps/plugin-dialog": "^2.4.0",
     "@tauri-apps/plugin-opener": "^2.5.0",
@@ -30,12 +31,15 @@
     "react-i18next": "^15.7.3",
     "tailwind-merge": "^3.6.0",
     "tailwindcss": "^4.1.18",
-    "tw-animate-css": "^1.4.0"
+    "tw-animate-css": "^1.4.0",
+    "zod": "^4.4.3",
+    "zustand": "^5.0.15"
   },
   "devDependencies": {
     "@rolldown/binding-linux-x64-gnu": "1.2.5",
     "@tailwindcss/oxide-linux-x64-gnu": "4.3.3",
     "@tailwindcss/vite": "^4.1.18",
+    "@tanstack/router-plugin": "^1.168.34",
     "@tauri-apps/cli": "^2.8.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.1.0",
