+++ apps/desktop/src/components/InsightsPage.tsx
@@ -141,7 +141,7 @@ function AchievementWall({ achievements }: { achievements: Achievement[] }) {
   const completedMilestones = tracks.reduce((count, item) => count + item.track.completed, 0);
   const milestoneCount = tracks.reduce((count, item) => count + item.track.milestones.length, 0);
   const completedSpecials = specials.filter((item) => item.unlocked).length;
-  return <Card className="overflow-hidden border-border shadow-sm"><CardHeader className="flex min-h-[64px] flex-row items-center justify-between gap-3 border-b border-border px-5 py-3"><div><h2 className="m-0 text-base font-semibold">{tr("insights.milestones")}</h2><p className="mt-0.5 text-xs text-muted-foreground">{completedMilestones} / {milestoneCount}</p></div><div className="flex items-center gap-2 max-[520px]:items-end max-[520px]:flex-col"><Badge variant="outline">{tr("achievementWall.milestones", { completed: completedMilestones, total: milestoneCount })}</Badge><Badge variant="outline">{tr("achievementWall.specials", { completed: completedSpecials, total: specials.length })}</Badge></div></CardHeader><CardContent className="p-0"><div className="grid grid-cols-[repeat(auto-fill,minmax(240px,1fr))] gap-4 bg-muted/20 p-5 max-[760px]:grid-cols-2 max-[520px]:grid-cols-1">{items.map((item) => <AchievementWallCard key={item.id} item={item} onOpen={() => setSelected(item)} />)}</div></CardContent>{selected && <AchievementDetailDialog key={selected.id} item={selected} onClose={() => setSelected(undefined)} />}</Card>;
+  return <Card className="overflow-hidden border-border shadow-sm"><CardHeader className="flex min-h-[64px] flex-row items-center justify-between gap-3 border-b border-border px-5 py-3"><div><div className="m-0 text-base font-semibold">{tr("insights.milestones")}</div><p className="mt-0.5 text-xs text-muted-foreground">{completedMilestones} / {milestoneCount}</p></div><div className="flex items-center gap-2 max-[520px]:items-end max-[520px]:flex-col"><Badge variant="outline">{tr("achievementWall.milestones", { completed: completedMilestones, total: milestoneCount })}</Badge><Badge variant="outline">{tr("achievementWall.specials", { completed: completedSpecials, total: specials.length })}</Badge></div></CardHeader><CardContent className="p-0"><div className="grid grid-cols-[repeat(auto-fill,minmax(240px,1fr))] gap-4 bg-muted/20 p-5 max-[760px]:grid-cols-2 max-[520px]:grid-cols-1">{items.map((item) => <AchievementWallCard key={item.id} item={item} onOpen={() => setSelected(item)} />)}</div></CardContent>{selected && <AchievementDetailDialog key={selected.id} item={selected} onClose={() => setSelected(undefined)} />}</Card>;
 }
 
 function AchievementWallCard({ item, onOpen }: { item: AchievementWallItem; onOpen: () => void }) {
