+++ apps/desktop/src/routes/agents.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/agents")({ component: () => null });
