+++ apps/desktop/src/routes/catalog.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/catalog")({ component: () => null });
