+++ apps/desktop/src/routes/insights.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/insights")({ component: () => null });
