+++ apps/desktop/src/routes/quota.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/quota")({ component: () => null });
