+++ apps/desktop/src/routes/settings.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/settings")({ component: () => null });
