+++ apps/desktop/src/routes/workspace/$workspaceId/context.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/workspace/$workspaceId/context")({ component: () => null });
