+++ apps/desktop/src/routes/workspace/$workspaceId/sessions.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/workspace/$workspaceId/sessions")({ component: () => null });
