+++ apps/desktop/src/routes/workspaces.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/workspaces")({ component: () => null });
